PROJECT1/pyproject.toml
[project]
name = "mlsolution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
]
